introduce literal
