+++ webknossos/webknossos/dataset/properties.py
@@ -8,6 +8,7 @@ import numpy as np
 import wkw
 from cattr.gen import make_dict_structure_fn, make_dict_unstructure_fn, override
 
+from webknossos.dataset.layer import LayerCategoryType
 from webknossos.geometry import BoundingBox, Mag
 from webknossos.utils import snake_to_camel_case
 
@@ -105,7 +106,7 @@ class MagViewProperties:
 @attr.define
 class LayerProperties:
     name: str
-    category: str
+    category: LayerCategoryType
     bounding_box: BoundingBox
     element_class: str
     wkw_resolutions: List[MagViewProperties]
