+++ wkcuber/wkcuber/api/layer_categories.py
@@ -1,0 +1,1 @@
+from webknossos.dataset.layer_categories import *  # pylint: disable=unused-wildcard-import,wildcard-import
